Fix handling for absolute tagfile paths
If a tagfile is not relative to the current directory, vim returns an
absolute path. `coc-lists` was assuming that all tagfiles would be
relative to the current working directory, and simply appended the
vim-returned paths to the current working directory path, resulting in
paths like `/home/user/project1/home/user/project2/tags`.

This checks for a leading `/` on returned tagfile-paths, and skips
prepending the current working directory if necessary.